+++ src/briefy/leica/views/assets.py
@@ -1,4 +1,5 @@
 """Views to handle Assets creation."""
+from briefy.ws.resources import BaseResource
 from briefy.ws.resources import RESTService
 from briefy.ws.resources import WorkflowAwareResource
 from briefy.leica.models import Asset
@@ -6,6 +7,8 @@ from briefy.leica.models import events
 from briefy.ws import CORS_POLICY
 from briefy.ws.resources.factory import BaseFactory
 from cornice.resource import resource
+from cornice.resource import view
+from pyramid.httpexceptions import HTTPNotFound as NotFound
 from pyramid.security import Allow
 
 
@@ -73,7 +76,58 @@ class AssetService(RESTService):
     factory=AssetFactory
 )
 class AssetWorkflow(WorkflowAwareResource):
-    """Assets workflow resourve."""
+    """Assets workflow resource."""
 
     model = Asset
     friendly_name = Asset.__name__
+
+
+@resource(
+    collection_path=PATH + '/versions',
+    path=PATH + '/versions/{version_id}',
+    cors_policy=CORS_POLICY,
+    factory=AssetFactory
+)
+class AssetVersions(BaseResource):
+    """Versioning of assets."""
+
+    model = Asset
+    friendly_name = Asset.__name__
+
+    @view(validators='_run_validators')
+    def collection_get(self):
+        """Return the list of versions for this object."""
+        id = self.request.matchdict.get('id', '')
+        obj = self.get_one(id)
+        raw_versions = obj.versions
+        versions = []
+        for version_id, version in enumerate(raw_versions):
+            versions.append(
+                {
+                    'id': version_id,
+                    'updated_at': version.updated_at
+                }
+            )
+        response = {
+            'versions': versions,
+            'total': obj.version + 1
+        }
+        return response
+
+    @view(validators='_run_validators')
+    def get(self):
+        """Return a version for this object."""
+        id = self.request.matchdict.get('id', '')
+        obj = self.get_one(id)
+        version_id = self.request.matchdict.get('version_id', 0)
+        try:
+            version_id = int(version_id)
+            version = obj.versions[version_id]
+        except (ValueError, IndexError):
+            raise NotFound(
+                '{friendly_name} with version: {id} not found.'.format(
+                    friendly_name=self.friendly_name,
+                    id=version_id
+                )
+            )
+        return version